Allow wildcard command blacklists and add command blacklists
This lets you blacklist all commands except those explicitly whitelisted, by blacklisting '*' and whitelisting the commands you want, fixing #48.

The whitelist is checked before the blacklist is checked (so if you blacklist and whitelist a command, you can still run it), and literals are prioritized over patterns (so a blacklisted '/tpa' overrides a whitelisted '/tp*').